loans8/requirements.txt
google-generativeai>=0.7.0
google-genai>=1.0.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
chromadb>=0.5.0
numpy<2
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
faster-whisper>=1.0.0
python-multipart>=0.0.9
